// asset-management/src/generated/models/Cr240_locationsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cr240_locationsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cr240_locationsstatecode = keyof typeof Cr240_locationsstatecode;
export const Cr240_locationsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cr240_locationsstatuscode = keyof typeof Cr240_locationsstatuscode;

export interface Cr240_locationsBase {
  cr240_locationid: string;
  cr240_locationname?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Cr240_locationsstatecode;
  statuscode?: Cr240_locationsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cr240_locations extends Cr240_locationsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}